Move inline tokens to same file
