--- packages/steamdown/src/lexer/inline-token.ts
@@ -1,5 +1,0 @@
-import Token from "./token";
-
-export default abstract class InlineToken extends Token {
-  public readonly scope = "inline";
-}
